Setup TypeScript Path Alias For New React App
Status Quo: In order to reference a module within the new react app, a
relative path must be used.

Problem: This often results in long chains of "parent directory" paths
(e.g. ../../../../../module-i-want).

Solution: Setup a path alias in the tsconfig file that maps to the src
directory of the react app. Then to reference a module within the app,
simply use the alias name and the path relative to the src directory
(e.g. @github-jobs/module-i-want).