Update Prettier To Use 72 Character Lines & No Semicolons
Status Quo: The default configuration for prettier is to insert
semicolons at the end of each statement and to use 82 characters wide
lines.

Problem: When using split screen on a 15 inch or 17 inch laptop, 82
characters-long lines tend to fold. Semicolons can add unnecessary
clutter to the code.

Solution: Update the prettier rc file to set the "semi" configuration
option to false, preventing the addition of semicolons. Set the
"printWidth" configuration option to 72 characters long.